Revert "update sdk"
This reverts commit 3a7c0a7248e9e47df38a72b0bb6d1e7b8050b992.

+++ esp/backend.c
@@ -70,7 +70,7 @@ void delay(uint16_t ms) {
 }
 void i2sTask(void*) {
     uint16_t samples = SAMPLES_PER_DURATION * SAMPLES_PER_BUFFER;
-    size_t bytesOut;
+    uint32_t bytesOut;
     while (true) {
         i2s_event_t event;
         // Wait indefinitely for a new message in the queue
